pyproject.toml
[tool.poetry]
name = "beta"
version = "0.1.0"
description = ""
authors = ["Your Name <youremail@domain.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-phonenumber-field = "^7.1.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.3"
phonenumbers = "^8.13.15"
rest-framework-simplejwt = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
